requirements.txt
# Core Framework
langchain==0.3.26 
angchain-core==0.3.70 
langchain-text-splitters==0.3.8
langchain-openai==0.3.28
langchain-naver==0.1.0
langgraph==0.5.4

# Data Processing
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.65

# Database
mysql-connector-python==8.4.0
pymysql==1.1.1
sqlalchemy==2.0.23

# API & Web
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0

# Utilities
python-dotenv==1.0.1
pydantic==2.7.4
typing-extensions==4.12.2
tqdm==4.67.1

# Scheduling (for daily updates)
schedule==1.2.0

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.21.1